backend/requirements.txt
fastapi
uvicorn
motor
pymongo
dnspython
python-dotenv
python-jose[cryptography]
bcrypt
pydantic
python-multipart
# sqlalchemy # REMOVED
